requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
azure-cognitiveservices-speech==1.32.1
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.4
instagrapi==2.0.0
jeepney==0.8.0
lz4==4.3.2
multidict==6.0.4
numpy==1.24.4
openai==0.28.1
pandas==2.0.3
Pillow==10.1.0
pycryptodomex==3.18.0
pydantic==1.10.9
pydub==0.25.1
pyktok==0.0.18
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
